requirements.txt
python==3.11
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1
scikit-learn==1.3.2
matplotlib==3.8.3
seaborn==0.13.2
PyPDF2==3.0.1
openai==1.12.0 